+++ spec/view_component/form/file_field_component_spec.rb
@@ -16,6 +16,28 @@ RSpec.describe ViewComponent::Form::FileFieldComponent, type: :component do
     end
   end
 
+  context "with direct upload" do
+    let(:options) { { direct_upload: true } }
+
+    it do
+      expect(component).to eq_html <<~HTML
+        <input data-direct-upload-url="http://test.host/rails/active_storage/direct_uploads" type="file" name="user[avatar]" id="user_avatar">
+      HTML
+    end
+  end
+
+  if Gem::Version.new(Rails::VERSION::STRING) >= Gem::Version.new("7.0")
+    context "with multiple and include hidden" do
+      let(:options) { { multiple: true, include_hidden: true } }
+
+      it do
+        expect(component).to eq_html <<~HTML
+          <input name="user[avatar][]" type="hidden" value="" autocomplete="off"><input multiple type="file" name="user[avatar][]" id="user_avatar">
+        HTML
+      end
+    end
+  end
+
   include_examples "component with custom html classes"
   include_examples "component with custom data attributes"
 end